Refactor roles modal
