+++ frontend/src/app/admin/layout.tsx
@@ -4,7 +4,7 @@ import Dashboard from "@/components/adminDashboard/adminDashboard";
 import Navbar from "@/components/navbar/navbar";
 import User from "@/types/user";
 import React, { Dispatch, SetStateAction } from "react";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 
 export const AdminContext = React.createContext<{
   users: User[];
@@ -16,14 +16,7 @@ export default function AdminLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const [name, setName] = useState<string>("");
   const [users, setUsers] = useState<User[]>([]);
-
-  useEffect(() => {
-    if (localStorage.getItem("name")) {
-      setName(String(localStorage.getItem("name")));
-    }
-  }, []);
 
   return (
     <AdminContext.Provider value={{ users, setUsers }}>
