+++ frontend/src/components/buttonsRow/buttonsRow.tsx
@@ -5,26 +5,15 @@ import styles from "./buttonsRow.module.css";
 import { AddButton, SubmitButton } from "@/components/button/button";
 import Modal from "@/components/modal/modal";
 import SearchDropdown from "../searchDropdown/searchDropdown";
-import { Options, SelectType } from "@/types/selectOptions";
-import { getDoorOptions } from "@/data/doors";
+import { SelectType } from "@/types/selectOptions";
 import axios from "@/axios/client";
 import User from "@/types/user";
 import { useRouter } from "next/navigation";
 
 export default function ButtonRow({ users }: { users: User[] }) {
   const [newKeyModal, setNewKeyModal] = useState(false);
-  const [newRoleModal, setNewRoleModal] = useState(false);
-  const [newRole, setNewRole] = useState<{ name: string }>({
-    name: "",
-  });
   const [selectedEmailOption, setSelectedEmailOption] =
     useState<SelectType>(null);
-  const [selectedRoleOption, setSelectedRoleOption] =
-    useState<SelectType>(null);
-  const [selectedDoorOption, setSelectedDoorOption] =
-    useState<SelectType>(null);
-  const [roles, setRoles] = useState<Options[]>([]);
-  const [doors, setDoors] = useState<Options[]>([]);
   const [disabled, setDisabled] = useState(false);
   const router = useRouter();
 
@@ -52,18 +41,14 @@ export default function ButtonRow({ users }: { users: User[] }) {
     {
       id: 3,
       name: "New Role",
-      onClick: async () => {
-        setDisabled(false);
-        setDoors(await getDoorOptions());
-        setNewRoleModal(true);
+      onClick: () => {
+        router.push("/admin/newRoleModal");
       },
     },
   ];
 
   const handleModalClose = () => {
     setNewKeyModal(false);
-    setNewRoleModal(false);
-    setNewRole({ name: "" });
     setSelectedEmailOption(null);
   };
 
@@ -75,31 +60,11 @@ export default function ButtonRow({ users }: { users: User[] }) {
         // @ts-ignore
         email: selectedEmailOption.label,
       });
-    } else if (newRoleModal) {
-      // handle adding new role
-      const rolesResp = await axios.post("/roles", {
-        name: newRole.name,
-        users: Array.isArray(selectedEmailOption)
-          ? selectedEmailOption.map((email: Options) => ({
-              id: parseInt(email.value),
-            }))
-          : [],
-        doors: Array.isArray(selectedDoorOption)
-          ? selectedDoorOption.map((door: Options) => ({
-              id: parseInt(door.value),
-            }))
-          : [],
-      });
-      const role = rolesResp.data;
-      setRoles([...roles, { label: role.name, value: role.id.toString() }]);
     }
 
     setSelectedEmailOption(null);
-    setSelectedDoorOption(null);
     setNewKeyModal(false);
-    setNewRoleModal(false);
     setDisabled(false);
-    setNewRole({ name: "" });
   };
 
   const newKeyModalContent = (
@@ -119,39 +84,6 @@ export default function ButtonRow({ users }: { users: User[] }) {
     </div>
   );
 
-  const newRoleModalContent = (
-    <div>
-      <h2 className={styles.subheader}>Name</h2>
-      <input
-        className={styles.modalInput}
-        type="text"
-        value={newRole.name}
-        onChange={(e) => setNewRole({ name: e.target.value })}
-      />
-      <SearchDropdown
-        options={emails}
-        placeholder="Add Emails..."
-        subheader="Emails"
-        selectDropdown="tableModal"
-        setSelectedOption={setSelectedEmailOption}
-        isMulti
-      />
-      <SearchDropdown
-        options={doors}
-        placeholder="Add Entrypoint..."
-        subheader="Entrypoints"
-        selectDropdown="tableModal"
-        setSelectedOption={setSelectedDoorOption}
-        isMulti
-      />
-      <SubmitButton
-        disabled={disabled}
-        text="Submit"
-        onClick={handleModalSubmit}
-      />
-    </div>
-  );
-
   return (
     <div className={styles.buttonRowContainer}>
       {buttons.map(({ id, name, onClick }) => (
@@ -162,13 +94,6 @@ export default function ButtonRow({ users }: { users: User[] }) {
           closeModal={handleModalClose}
           title="New Key"
           content={newKeyModalContent}
-        />
-      )}
-      {newRoleModal && (
-        <Modal
-          closeModal={handleModalClose}
-          title="New Role"
-          content={newRoleModalContent}
         />
       )}
     </div>
